Last project update (i think)

+++ MySQLDB.py
@@ -1,5 +1,6 @@
 import mysql.connector
 import random
+import settings
 
 dbconfig = {'host': 'eu-cdbr-west-02.cleardb.net',
             'user': 'bca611917ce10f',
@@ -7,7 +8,7 @@ dbconfig = {'host': 'eu-cdbr-west-02.cleardb.net',
             'database': 'heroku_4e4d24d83d2a5ae'}
 
 
-def add_user(name, id):
+def add_user(name, id, url):
     answer = ""
     conn = mysql.connector.connect(**dbconfig)
     cur = conn.cursor()
@@ -17,7 +18,7 @@ def add_user(name, id):
     if len(res) > 0:
         answer = False
     else:
-        _SQL = f'INSERT INTO users(name, id) VALUES("{str(name)}", {str(id)});'
+        _SQL = f'INSERT INTO users(name, id, image_url) VALUES("{str(name)}", {str(id)}, "{str(url)}");'
         cur.execute(_SQL)
         conn.commit()
         answer = True
@@ -32,17 +33,20 @@ def registration(id, match_id, civ1, civ2):
     """
     conn = mysql.connector.connect(**dbconfig)
     cur = conn.cursor()
-    _SQL = f'SELECT id FROM users WHERE id={str(id)}' # Получаем юзера из БД
+    _SQL = f'SELECT id FROM users WHERE id={str(id)}'  # Получаем юзера из БД
     cur.execute(_SQL)
     res = cur.fetchall()
     if len(res) == 0:
         cur.close()
         conn.close()
-        return False, False, False #Есть в БД? Может участвовать в матче? Может присоединиться к ЭТОМУ матчу?
-    _SQL = f'SELECT * FROM status WHERE user_id = {str(id)}' #Получаем информацию о нахождении в игре
+        # Есть в БД? Может участвовать в матче? Может присоединиться к ЭТОМУ
+        # матчу?
+        return False, False, False
+    # Получаем информацию о нахождении в игре
+    _SQL = f'SELECT * FROM status WHERE user_id = {str(id)}'
     cur.execute(_SQL)
     res = cur.fetchall()
-    if len(res) == 0: #Если не в игре, то идём дальше
+    if len(res) == 0:  # Если не в игре, то идём дальше
         _SQL = f'SELECT start FROM match_status WHERE match_id = {str(match_id)}'
         cur.execute(_SQL)
         res = cur.fetchone()
@@ -50,7 +54,9 @@ def registration(id, match_id, civ1, civ2):
             if res[0] == 1:
                 cur.close()
                 conn.close()
-                return True, True, False #Пользователь не в игре, НО регистрация на текущий матч закрыта
+                # Пользователь не в игре, НО регистрация на текущий матч
+                # закрыта
+                return True, True, False
             else:
                 _SQL = f'INSERT INTO status(match_id, user_id) VALUES({str(match_id)}, {str(id)})'
                 cur.execute(_SQL)
@@ -59,7 +65,7 @@ def registration(id, match_id, civ1, civ2):
                 conn.commit()
                 cur.close()
                 conn.close()
-                return True, True, True #Пользователь присоединился к этой игре
+                return True, True, True  # Пользователь присоединился к этой игре
         except TypeError:
             _SQL = f'INSERT INTO match_status(match_id) VALUES({str(match_id)})'
             cur.execute(_SQL)
@@ -70,20 +76,21 @@ def registration(id, match_id, civ1, civ2):
             conn.commit()
             cur.close()
             conn.close()
-            return True, True, True #Пользователь создал новую игру.
+            return True, True, True  # Пользователь создал новую игру.
     else:
-        return True, False, False #В базе данных, но находится в игре.
+        return True, False, False  # В базе данных, но находится в игре.
+
 
 def leave(id):
     conn = mysql.connector.connect(**dbconfig)
     cur = conn.cursor()
-    _SQL = f'SELECT id FROM users WHERE id={str(id)}' # Получаем юзера из БД
+    _SQL = f'SELECT id FROM users WHERE id={str(id)}'  # Получаем юзера из БД
     cur.execute(_SQL)
     res = cur.fetchall()
     if len(res) == 0:
         cur.close()
         conn.close()
-        return False, False, False #Нет в БД
+        return False, False, False  # Нет в БД
     try:
         _SQL = f'SELECT match_id FROM status WHERE user_id = {str(id)}'
         cur.execute(_SQL)
@@ -94,14 +101,14 @@ def leave(id):
     except TypeError:
         return True, False, False
     if res2[0] == 1:
-        return True, True, False  #Матч уже начался
+        return True, True, False  # Матч уже начался
     else:
         _SQL = f'DELETE FROM status WHERE user_id = {str(id)}'
         cur.execute(_SQL)
         conn.commit()
         cur.close()
         conn.close()
-        return True, True, True #Вышел из матча
+        return True, True, True  # Вышел из матча
 
 
 def unbanned(id):
@@ -112,27 +119,29 @@ def unbanned(id):
     _SQL = f'DELETE FROM banned WHERE user_id={str(id)}'
     cur.execute(_SQL)
     banned_civs = cur.fetchone()
+    cur.close()
+    conn.close()
     return banned_civs[0], banned_civs[1]
 
 
-def host_rand(players):
+def host_rand():
     sum = 0
-    host_index = random.randint(0, len(players))
+    host_index = random.randint(0, len(settings.users_id) - 1)
     conn = mysql.connector.connect(**dbconfig)
     cur = conn.cursor()
-    _SQL = f'SELECT match_id FROM status WHERE user_id={str(players[0])}'
+    _SQL = f'SELECT match_id FROM status WHERE user_id={str(settings.users_id[0])}'
     cur.execute(_SQL)
     res = cur.fetchone()
-    for player in players:
-        _SQL = f'SELECT point FROM users WHERE id={str(player)}'
+    for player in settings.users_id:
+        _SQL = f'SELECT points FROM users WHERE id={str(player)}'
         cur.execute(_SQL)
         points = cur.fetchone()
         sum += points[0]
-    avg = sum/len(players)
+    avg = sum / len(settings.users_id)
     _SQL = f'UPDATE match_status SET average_points={str(avg)} WHERE match_id={str(res[0])}'
     cur.execute(_SQL)
     conn.commit()
-    _SQL = f'UPDATE match_status SET host_id={str(players[host_index])} WHERE match_id={str(res[0])}'
+    _SQL = f'UPDATE match_status SET host_id={str(settings.users_id[host_index])} WHERE match_id={str(res[0])}'
     cur.execute(_SQL)
     conn.commit()
     cur.close()
@@ -145,12 +154,24 @@ def start(id):
     _SQL = f'UPDATE match_status SET start=1 WHERE match_id={str(id)}'
     cur.execute(_SQL)
     conn.commit()
-    _SQL = f'SELECT host_id, average_points WHERE match_id={str(id)}'
+    _SQL = f'SELECT average_points FROM match_status WHERE match_id={str(id)}'
     cur.execute(_SQL)
-    res = cur.fetchone()
+    avg_points = cur.fetchone()
+    _SQL = f'SELECT host_id FROM match_status WHERE match_id={str(id)}'
+    cur.execute(_SQL)
+    host_id = cur.fetchone()
+    _SQL = f'SELECT name FROM users WHERE id={str(host_id[0])}'
+    cur.execute(_SQL)
+    host_name = cur.fetchone()
+    _SQL = f'SELECT image_url FROM users WHERE id={str(host_id[0])}'
+    cur.execute(_SQL)
+    host_avatar = cur.fetchone()
+    _SQL = f'DELETE FROM banned'
+    cur.execute(_SQL)
+    conn.commit()
     cur.close()
     conn.close()
-    return res
+    return avg_points[0], host_name[0], host_avatar[0]
 
 
 def statistic(id):
@@ -167,14 +188,14 @@ def statistic(id):
 def result(host, id, *args):
     conn = mysql.connector.connect(**dbconfig)
     cur = conn.cursor()
-    _SQL = f'SELECT host_id from match_status WHERE match_id={str(id)}'
+    _SQL = f'SELECT host_id FROM match_status WHERE match_id={str(id)}'
     cur.execute(_SQL)
     res = cur.fetchone()
     if res[0] != host:
         return False
     else:
-        for player in enumerate(args):
-            _SQL = f'SELECT id FROM users WHERE name={str(player[1])}'
+        for player in enumerate(*args):
+            _SQL = f'SELECT id FROM users WHERE name="{str(player[1])}"'
             cur.execute(_SQL)
             res = cur.fetchone()
             _SQL = f'DELETE FROM status WHERE user_id={str(res[0])}'
@@ -182,7 +203,7 @@ def result(host, id, *args):
             conn.commit()
             _SQL = f'SELECT average_points FROM match_status WHERE host_id={str(host)}'
             cur.execute(_SQL)
-            avg_points = cur.fetchone()[0]
+            avg_points = cur.fetchone()
             _SQL = f'UPDATE users SET games = games + 1 WHERE id={str(res[0])}'
             cur.execute(_SQL)
             conn.commit()
@@ -190,30 +211,36 @@ def result(host, id, *args):
                 _SQL = f'UPDATE users SET wins = wins + 1 WHERE id={str(res[0])}'
                 cur.execute(_SQL)
                 conn.commit()
-                _SQL = f'UPDATE users SET points = points + {int(avg_points*0.35)} WHERE id={str(res[0])}'
+                _SQL = f'UPDATE users SET points = points + {int(int(avg_points[0])*0.35)} WHERE id={str(res[0])}'
                 cur.execute(_SQL)
                 conn.commit()
             else:
                 _SQL = f'UPDATE users SET loses = loses + 1 WHERE id={str(res[0])}'
                 cur.execute(_SQL)
                 conn.commit()
-                _SQL = f'UPDATE users SET points = points - {int(avg_points*0.15)} WHERE id={str(res[0])}'
+                _SQL = f'UPDATE users SET points = points - {int(int(avg_points[0])*0.15)} WHERE id={str(res[0])}'
                 cur.execute(_SQL)
                 conn.commit()
         _SQL = f'SELECT user_id FROM status WHERE match_id={str(id)}'
         cur.execute(_SQL)
         res = cur.fetchall()
-        for survivors in res:
-            _SQL = f'DELETE FROM status WHERE user_id={str(survivors[0])}'
-            cur.execute(_SQL)
-            conn.commit()
-            _SQL = f'UPDATE users SET games = games + 1 WHERE id={str(survivors[0])}'
-            cur.execute(_SQL)
-            conn.commit()
-            _SQL = f'UPDATE users SET survives = survives + 1 WHERE id={str(survivors[0])}'
-            cur.execute(_SQL)
-            conn.commit()
-            _SQL = f'UPDATE users SET points = points + {int(avg_points*0.2)} WHERE id={str(survivors[0])}'
+        if len(res) != 0:
+            for survivors in res:
+                _SQL = f'DELETE FROM status WHERE user_id={str(survivors[0])}'
+                cur.execute(_SQL)
+                conn.commit()
+                _SQL = f'UPDATE users SET games = games + 1 WHERE id={str(survivors[0])}'
+                cur.execute(_SQL)
+                conn.commit()
+                _SQL = f'UPDATE users SET survives = survives + 1 WHERE id={str(survivors[0])}'
+                cur.execute(_SQL)
+                conn.commit()
+                _SQL = f'UPDATE users SET points = points + {int(int(avg_points[0])*0.2)} WHERE id={str(survivors[0])}'
+                cur.execute(_SQL)
+                conn.commit()
+        _SQL = f'DELETE FROM status WHERE match_id={str(id)}'
+        cur.execute()
+        conn.commit()
         cur.close()
         conn.close()
         return True
